+++ pipeline/src/pipeline/api/mod.rs
@@ -11,5 +11,6 @@ pub mod step_dependency;
 pub mod step_list;
 pub mod step_new;
 pub mod step_output;
+pub mod step_remove;
 pub mod step_show;
 pub mod step_update;