+++ pipeline/src/pipeline/step.rs
@@ -14,6 +14,7 @@ use xvc_ecs::{persist, XvcEntity};
 use xvc_logging::XvcOutputSender;
 
 use super::api::step_list::cmd_step_list;
+use super::api::step_remove::cmd_step_remove;
 use super::XvcStepInvalidate;
 
 /// Step creation, dependency, output commands
@@ -54,6 +55,14 @@ pub enum StepSubCommand {
         when: Option<XvcStepInvalidate>,
     },
 
+    /// Remove a step from a pipeline
+    #[command()]
+    Remove {
+        /// Name of the step to remove
+        #[arg(long, short)]
+        step_name: String,
+    },
+
     /// Update a step's command or when options.
     #[command(about = "Update step options")]
     Update {
@@ -75,7 +84,7 @@ pub enum StepSubCommand {
     #[command()]
     Dependency {
         /// Name of the step to add the dependency to
-        #[arg(long, short)]
+        #[arg(long, short, aliases = &["for", "to"])]
         step_name: String,
 
         /// Add a generic command output as a dependency. Can be used multiple times.
@@ -218,6 +227,10 @@ pub fn handle_step_cli(
             command,
             when: changed,
         } => cmd_step_new(xvc_root, pipeline_name, step_name, command, changed),
+
+        StepSubCommand::Remove { step_name } => {
+            cmd_step_remove(output_snd, xvc_root, pipeline_name, step_name)
+        }
 
         StepSubCommand::Update {
             step_name,