+++ workflow_tests/tests/test_pipeline_dag.rs
@@ -1,7 +1,5 @@
 mod common;
 
-use std::{fs, path::Path};
-
 use common::*;
 use log::LevelFilter;
 use xvc::error::Result;